Added some initial CSS
