+++ langtastic/src/components/Dashboard/BookingInformation/TutorInformation/TutorLanguages/index.css
@@ -1,0 +1,3 @@
+.flex {
+  display: flex;
+}
