+++ langtastic/src/components/Dashboard/BookingInformation/TutorInformation/TutorLanguages/index.js
@@ -1,10 +1,11 @@
+import "./index.css";
 function TutorLanguages({ languagesArray }) {
   return (
-    <>
+    <div className="flex">
       {languagesArray.map((item, index) => {
         return <p key={index}>{item}</p>;
       })}
-    </>
+    </div>
   );
 }
 export default TutorLanguages;
