+++ langtastic/src/components/Dashboard/BookingInformation/TutorInformation/TutorPicture/index.css
@@ -1,4 +1,6 @@
 img {
   border: solid 2px;
   border-radius: 100%;
+  width: 100px;
+  height: 100px;
 }
